[asl] [tool] Build a big file with instructions
This commit introduces a new option, `-m` or `--make-functions`, which
process all the files passed in input, and for all instruction files,
write a corresponding function in the output file `instructions.asl`.

Compared to previous features of this script, this feature build ASL
code corresponding to the decode part of the instruction. Indeed, the
built function has only take a `bits(32)` type as input: the
instruction in binary format.

We currently extract only one decode section per instruction.