[herd,aarch64,asl] No ASL evaluation for some 64bit comparisons
Namely it is not possible to let ASL code operate when
comparison applies to non-scalars, such as symbolic addresses
or PTE values. Indeed, comparison is an alias for subtraction
to register ZR, implemented by logical complement followed
by add with carry set.

In order to follow the ASL path as much as possible
the bypass applies to pure register to register and register
to zero comparisons only.